Folds DimmWitted class into dimmwitted.{cc,h}
