+++ Makefile
@@ -52,7 +52,6 @@ SOURCES += src/variable.cc
 SOURCES += src/factor.cc
 SOURCES += src/factor_graph.cc
 SOURCES += src/inference_result.cc
-SOURCES += src/gibbs_sampling.cc
 SOURCES += src/gibbs_sampler.cc
 SOURCES += src/timer.cc
 OBJECTS = $(SOURCES:.cc=.o)
